feat: #108454 公开课课时是否可编辑

+++ src/ApiBundle/Api/Resource/OpenCourse/OpenCourseLesson.php
@@ -49,6 +49,9 @@ class OpenCourseLesson extends AbstractResource
         if (!empty($lesson) && $courseId != $lesson['courseId']) {
             throw CommonException::ERROR_PARAMETER();
         }
+        if ('published' == $lesson['status']) {
+            throw OpenCourseException::DELETE_PUBLISHED_LESSON();
+        }
         $this->getOpenCourseService()->deleteLesson($lessonId);
 
         return ['ok' => true];
